# Optimize push() helper in parse_expr()