Requires #158, #163.

```
------------------------------------------------------------------------------------
Benchmark                          Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------
fizzy/parse/mul256_opt0         3177 ns         3177 ns      1326707 rate=263.804M/s size=838
fizzy/parse/micro/spinner        610 ns          610 ns      6836433 rate=104.972M/s size=64
Comparing bin/fizzy-bench-master to bin/fizzy-bench
Benchmark                            Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------
fizzy/parse/mul256_opt0           -0.2923         -0.2923          4488          3177          4488          3177
fizzy/parse/micro/spinner         +0.0298         +0.0298           592           610           592           610
```