@chfast 2020-02-21
> Why is there both `push` and `store` in this file which seemingly do very similar things?

Still they are doing something different.