@gumb0 2020-02-24
Why can it be faster?