@chfast 2020-02-24
> Why can it be faster?

In the prev version `.resize()` is used to extend the string. This calls the `memset()` to zero the extension. This call is definitely not needed and quite expensive in the same time.